Add data types, parsers to support the full numeric tower of Scheme.
https://en.wikibooks.org/wiki/Write_Yourself_a_Scheme_in_48_Hours/Parsing

Add data types and parsers to support the full numeric tower of Scheme
numeric types. Haskell has built-in types to represent many of these; check
the Prelude. For the others, you can define compound types that represent
eg. a Rational as a numerator and denominator, or a Complex as a real and
imaginary part (each itself a Real).